backend/requirements.txt
fastapi==0.109.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
python-dotenv==1.0.1
python-multipart==0.0.6
scikit-learn==1.4.2
scipy==1.11.4
seaborn==0.13.1
statsmodels==0.14.2
uvicorn[standard]==0.27.0